+++ src/components/section-fronts/SectionFrontPage/SectionFrontPage.tsx
@@ -34,6 +34,7 @@ export const SECTION_FRONT_QUERY = gql`
         ... on Donation {
           __typename
           id
+          status
           type
           title
           description
@@ -56,6 +57,7 @@ export const SECTION_FRONT_QUERY = gql`
         ... on Jumbotron {
           __typename
           id
+          status
           type
           title
           description
@@ -94,6 +96,7 @@ export const SECTION_FRONT_QUERY = gql`
         ... on CardGrid {
           __typename
           id
+          status
           type
           title
           description
@@ -123,6 +126,7 @@ export const SECTION_FRONT_QUERY = gql`
         ... on ExternalSearch {
           __typename
           id
+          status
           type
           title
           description
@@ -135,6 +139,7 @@ export const SECTION_FRONT_QUERY = gql`
         ... on EmailSubscription {
           __typename
           id
+          status
           type
           heading
           description
@@ -148,6 +153,7 @@ export const SECTION_FRONT_QUERY = gql`
         ... on ButtonLinks {
           __typename
           id
+          status
           type
           heading
           description
@@ -164,6 +170,7 @@ export const SECTION_FRONT_QUERY = gql`
         ... on Text {
           __typename
           id
+          status
           type
           heading
           text